correction to prev

+++ formative/signals.py
@@ -182,26 +182,51 @@ def collectionblock_post_save(sender, instance, created, raw, **kwargs):
                                       page=0, _rank__gt=1).delete()
     
     text, style = default_text(block.name) + ':', FormLabel.LabelStyle.VERTICAL
-    paths = []
-    l = FormLabel.objects.get_or_create(form=block.form, path=block.name,
-                                        defaults={'style': style, 'text': text})
-    paths.append(l[0].path)
+    paths, id_name, l = [], f'{block.name}{block.id}', None
+    try: l = FormLabel.objects.get(form=block.form, path=id_name+'_')
+    except FormLabel.DoesNotExist:
+        l, _ = FormLabel.objects.get_or_create(form=block.form, path=block.name,
+                                               defaults={'style': style,
+                                                         'text': text})
+    paths.append(l.path)
     
     if block.fixed: return # no implementation yet for fixed + collection
     style = FormLabel.LabelStyle.WIDGET
     for name in block.collection_fields():
-        text, path = default_text(name), '.'.join((block.name, name))
-        l = FormLabel.objects.get_or_create(form=block.form, path=path,
-                                            defaults={'style': style,
-                                                      'text': text})
-        paths.append(l[0].path)
+        text = default_text(name)
+        path, id_path = '.'.join((block.name, name)), id_name + f'.{name}_'
+        try: l = FormLabel.objects.get(form=block.form, path=id_path)
+        except FormLabel.DoesNotExist:
+            l, _ = FormLabel.objects.get_or_create(form=block.form, path=path,
+                                                   defaults={'style': style,
+                                                             'text': text})
+        paths.append(l.path)
     
-    delete_sub_labels(block.form, block.name, paths)
+    # TODO: field name change: update name of existing label rather than delete
+    
+    delete_sub_labels(block.form, id_name, paths, True)
     if not new and block._old_name != block.name:
-        delete_sub_labels(block.form, block._old_name)
+        delete_sub_labels(block.form, f'{block._old_name}{block.id}', True)
+    
+    labels = block.form.labels
+    blocks = block.form.collections(name=block.name).exclude(id=block.pk)
+    if not blocks: delete_sub_labels(block.form, block.name, paths)
+    else:
+        for label in labels.filter(path__startswith=block.name+'.'):
+            if label.path in paths: continue
+            field = label.path[len(block.name)+1:]
+            if not blocks.filter(any_name_field(_=field)): label.delete()
+    
+    if not new and block._old_name != block.name:
+        blocks = block.form.collections(name=block._old_name)
+        if not blocks: delete_sub_labels(block.form, block._old_name)
+        else:
+            for label in labels.filter(path__startswith=block._old_name+'.'):
+                if not blocks.filter(any_name_field(_=field)): label.delete()
 
-def delete_sub_labels(form, name, exclude=[]):
+def delete_sub_labels(form, name, exclude=[], id=False):
     sl = form.labels.filter(path__startswith=name+'.').exclude(path__in=exclude)
+    if id: sl = sl.filter(path__endswith='_')
     sl.delete()
 
 def delete_block_labels(form, name):
